test(jarm): rename stale test method to match ToDeliveryMode
