+++ Abblix.Oidc.Server.Mvc/Formatters/AuthorizationErrorFormatter.cs
@@ -24,6 +24,7 @@ using Abblix.Oidc.Server.Common.Constants;
 using Abblix.Oidc.Server.Endpoints.Authorization.Interfaces;
 using Abblix.Oidc.Server.Endpoints.Configuration.Interfaces;
 using Abblix.Oidc.Server.Features.Issuer;
+using Abblix.Oidc.Server.Features.Jarm;
 using Abblix.Oidc.Server.Features.Tokens.Formatters;
 using Abblix.Oidc.Server.Model;
 using Abblix.Oidc.Server.Mvc.Binders;
@@ -90,7 +91,7 @@ public class AuthorizationErrorFormatter(
                         responseType.Contains(ResponseTypes.Token) ||
                         responseType.Contains(ResponseTypes.IdToken));
 
-                    responseMode = responseMode.ToDeliveryMode(carriesTokens);
+                    responseMode = ResponseModes.ToDeliveryMode(responseMode, carriesTokens);
                 }
 
                 return await FormatResponseAsync(response, responseMode, redirectUri);
